docs: Update the lint settings. (#53)
The lint command already specified in the package.json appears to not
have been run for quite a while, ran it and fixed a number of the
issues.  There were a few actual code fixes that would cause breakage in
a compliant browser - const url = ... and then updating the actual URL
was changed to let url = ... to avoid that defect.  However, the build
command removed the "const" part of this, so it wasn't ever being seen
as a real issue.