feat(config): add quoteAcceptHeaderParameterValues as config.
According to DICOM Part 3.18, the quotes are required and
some origin servers fail if they are not provided. However, the original RFC 2387 does not require the quotes.
Most servers can handle no quotation accept header and including the quotation for them will trigger a preflight
request and becomes a performance issue (a preflight request is sent out for each request = 2 times the amount
of requests over the network) An in length discussion can be read here https://github.com/dcmjs-org/dicomweb-client/issues/57)